requirements.txt
Package                    Version
-------------------------- -------
pip                        24.3.1
setuptools                 66.1.1
TestLink-API-Python-client 0.8.1
wheel                      0.38.4
